Marc made some major improvements to the velocity finding algorithm.
Gaussian low pass filter on the curvdata itself (spatially, not temporally)
Accumulated averaging (exponential decay) with one weight, alpha at .85
Instead of cropping 10% from head tail, we are inducing an assymetry and cropping 25% of head and 5% of tail.
Smoothing the final phase velocity with lowpass1d.

Things to do tomorrow. Go through and comment. Have an options structure. Also flag which of marc's functions are required and pull those.